requirements.txt
aenum==2.2.3
amqp==2.5.2
asgiref==3.3.1
Babel==2.8.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
boto3==1.12.25
botocore==1.15.25
celery==4.4.2
certifi==2020.4.5.1
cffi
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
DAWG==0.8.0
DAWG-Python==0.7.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.4.0
Django==3.1.7
django-admin-autocomplete-filter==0.6.1
django-autocomplete-light==3.8.2
django-bootstrap-form==3.4
django-bootstrap3==14.2.0
django-countries==6.1
django-debug-toolbar==3.2
django-import-export==2.5.0
django-markdownx==3.0.1
django-paypal==1.0.0
django-phonenumber-field==4.0.0
django-recaptcha==2.0.6
django-storages==1.9.1
# https://github.com/darklow/django-suit/tarball/v2
# django-suit==2.0a1
django-telegrambot==1.0.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
emoji==0.5.4
et-xmlfile==1.0.1
facebook-sdk==3.1.0
factory-boy==2.12.0
Faker==4.1.0
future==0.18.2
idna==2.9
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.9.5
kombu==4.6.8
lockfile==0.12.2
mailchimp3==3.0.13
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mock==4.0.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.4
pandas
pdfkit==0.6.1
phonenumbers==8.12.0
pilkit==2.0
Pillow
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.4.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
qrcode==6.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
text-unidecode==1.3
tornado==6.0.4
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
uWSGI==2.0.18
vine==1.3.0
wikipedia==1.4.0
xlrd==1.2.0
xlwt==1.3.0
yookassa>=2.3.1

gunicorn
gevent

sentry-sdk==0.14.3

stripe>=2.48.0,<3.0
drf-api-tracking==1.7.8

user-agents==2.2.0
